Refactor directories
Utility functions with no sideeffects moved to utils

Project creators and sccafolders stay in helpers
