+++ src/helpers/logNextSteps.ts
@@ -1,11 +1,11 @@
 import type { Packages } from "../index";
 import chalk from "chalk";
-import { getPkgManager } from "./get-pkg-manager";
-import { logger } from "./logger";
+import { getUserPkgManager } from "../utils/getUserPkgManager";
+import { logger } from "../utils/logger";
 
 // This logs the next steps that the user should take in order to advance the project
 export const logNextSteps = (projectName: string, packages: Packages) => {
-  const pkgManager = getPkgManager();
+  const pkgManager = getUserPkgManager();
 
   logger.info("Next steps:");
   logger.info(` cd ${chalk.cyan.bold(projectName)}`);
